crysgarage-admin/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.6
pydantic==2.7.4
pydantic-settings==2.3.4
SQLAlchemy==2.0.32
alembic==1.13.2
asyncpg==0.29.0
aiosqlite==0.20.0
httpx==0.27.2
itsdangerous==2.2.0
PyJWT==2.9.0
firebase-admin==6.5.0
python-dotenv==1.0.1
pytest==8.3.2
pytest-asyncio==0.23.8
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
